[examples] Finish breakout game logic (#6)
This commit sets the game up to draw all of the boxes in the breakout
game, as well as the paddle and ball. It also gives collision and motion
logic to the ball. The game logic is mostly untested, but it's just an
example of the I2C library - all the important code is in the oled
display code.